requirements.txt
discord.py==0.16.12
python-dotenv==0.10.1 
requests==2.21.0
schedule==0.6.0
beautifulsoup4==4.7.1